コントローラー編集・requestspec導入
