+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-  resources :categories do
-    resources :ideas
+  namespace 'api' do
+    namespace 'v1' do
+      resources :categories, only: [:index, :create]
+    end
   end
 end
